hitl-pdf-preview/server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pdfminer.six==20231228
ebooklib==0.18
lxml==5.3.0
Pillow==11.0.0
openpyxl==3.1.5
xlrd==2.0.1
openai>=1.30.0
python-dotenv>=1.0.0
